[vm] Lower smi value used for setting identity hash codes
In 64-bit with compressed pointers enabled, we still store identity
hashes in the object header, but the Smi range is reduced.

The 0xffffffff is a Smi in 64-bit architectures, but not if
compressed-pointers are enabled.

This updates the test to use the same maximum value as the corelibraries
could use when setting the identity hash code.

TEST=Fixes vm/cc/AsmIntrinsifier_SetHashIfNotSetYet in compressed-pointers mode

Change-Id: I285a28127767a6a0b606354f3e3f994a15b372b5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/202480
Reviewed-by: Ryan Macnak <rmacnak@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>