Read pubspec.yaml file once, during Workspace creation.
This makes it less probably that we will get inconsistent results,
such as computing the salt with one file content, but then read a
different content for parsing and reporting lints.

But I think there there is still something not quite correct with
deciding that there was a change to pubspec.yaml, we might miss
some events.

Bug: https://github.com/dart-lang/linter/issues/2694
Change-Id: Ied76dc2f6c1203d3eed68e5220196d91c6d8e717
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/202726
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>