Create a typedef for list-of-non-promotion-reasons
I think the signature of this type is long and complex enough to warrant
a typedef. It is both complex (nested type parameters, and a function
type) and long (`List<Map<DartType, NonPromotionReason> Function()>` is
50 characters by itself).

This does require bumping the language version of analyzer to 2.13.

Change-Id: I467f6ec800acd0a41d9bec429e22f85020b1543e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/203641
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>