[parser] Different recovery for symbols with non-identifiers
Instead of getting errors like this:

```
t.dart:6:4: Error: Expected an identifier, but got 'class'.
Try inserting an identifier before 'class'.
  #class.foo;
   ^^^^^
t.dart:6:4: Error: Expected ';' after this.
  #class.foo;
   ^^^^^
t.dart:6:4: Error: 'class' can't be used as an identifier because it's a keyword.
Try renaming this to be an identifier that isn't a keyword.
  #class.foo;
   ^^^^^
t.dart:6:4: Error: Getter not found: 'class'.
  #class.foo;
   ^^^^^
```

we get something like this:

```
t.dart:6:4: Error: 'class' can't be used as an identifier because it's a keyword.
Try renaming this to be an identifier that isn't a keyword.
  #class.foo;
   ^^^^^
```

https://github.com/dart-lang/sdk/issues/46208
https://github.com/dart-lang/sdk/issues/46207

Change-Id: I79da73e600f732d10020c871d11e11c3f7fe740a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/203686
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>