FunctionType.element should be null for aliasedType.
This used to be the case when we read element model from bytes.
But we started (accidentally) providing it when keeping linking elements.
This caused https://dart-review.googlesource.com/c/sdk/+/201661 to be
reverted.

There is a work around in dartdoc, to be pulled into SDK.
https://github.com/dart-lang/dartdoc/pull/2687

This CL hopefully fixes the issue.

Change-Id: Ib1a952085a80c4673baab01a6c4e4a30c1a8ea73
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/204360
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>