[vm] Create macros for wrapping non-ObjectPtr fields with WSRs.
The new macros mean that the field type is only generalized to ObjectPtr
in the precompiler, and otherwise stays the same. Since the serializer
is the only part of the system that needs to know about
WeakSerializationReferences and it accesses the fields of the untagged
object directly, have the macro create a getter that automatically
unwraps if needed and casts the wrapped value to the normally expected
type so no other code in the VM needs to change.

This change also changes WeakSerializationReference::New to return an
ObjectPtr, not a WeakSerializationReferencePtr, so we can return the
target instead of wrapping it if it's an object that's guaranteed to be
serialized (e.g., null).

WeakSerializationReference::New also checks to see if the target is
already a WeakSerializationReference, and returns it if the replacement
is the same. If the replacement is different, it uses the target of the
WSR as its target.  This way, WSRs are never nested, which fits the
assumptions from the original implementation.

TEST=Primarily a refactoring, so existing tests.

Cq-Include-Trybots: luci.dart.try:vm-kernel-linux-debug-x64-try,vm-kernel-nnbd-linux-debug-x64-try,vm-kernel-precomp-linux-debug-x64-try,vm-kernel-precomp-linux-debug-x64c-try,vm-kernel-precomp-nnbd-linux-debug-x64-try,vm-kernel-linux-debug-x64c-try
Change-Id: Ia6790e4e5bd5cf302b732180dee42f796d28d59e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/204141
Commit-Queue: Tess Strickland <sstrickl@google.com>
Reviewed-by: Daco Harkes <dacoharkes@google.com>
Reviewed-by: Ryan Macnak <rmacnak@google.com>