Hints etc. to prep for migrating the migration tool
This change adds just enough migration hints so that I can run the
migration tool on itself and get reasonable results.  Most of the
hints are just marking a few types as nullable or non-nullable that
the migration tool isn't able to correctly infer by itself.  In a few
circumstances I've had to make a type explicit that was previously
inferred, so that I could mark it as nullable.  In a few other
circumstances I've had to add `/*!*/` after an expression to let the
migration tool know that the expression should be null checked.

Note that unlike most of the rest of the Dart SDK, the migration tool
is not expected to be a long term part of the Dart ecosystem, since
the whole point of it is to help people migrate their code to null
safety.  It's not likely to receive a lot more maintenance, and it
will likely be removed sometime around the time of Dart 3.0 once
people stop needing it.  Accordingly, I've tried to err on the side of
liberally allowing types to be nullable where possible.  Although this
means that we will wind up with more null checks in the final result,
it makes the migration process a lot easier and substantially reduces
the risk of introducing bugs during the migration process.

Change-Id: Ia348a09e94d22d83c2c15215ca7ddf5a5a4d05f9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/204480
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>