[CFE] Fix weekly tests didn't wait correctly
When running the weekly test it launches a leak test for flutter first
without the experimental invalidation and then with. As the test starts
with checking out flutter gallery the second launch has to wait for the
first to be done with that.
That stopped working with 8bd4c533812138750d3fc175ff807c9d0921f496 as
the (in that CL) newly added test trickered that wait to stop (by
printing a "globally listened for" observatory listening on line).

This CL fixes that by only listening for data from the correct process.

It further more makes final error-reporting better.

Change-Id: I1ead96ae4c31616b1e7dc3044ae22437d7f605d5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/204783
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>