Clean up annotation handling in ScopedVisitor and ResolverVisitor.
We need to be careful that when analyzing a declaration that
introduces a new scope, any annotations on that declaration are
considered to be outside of the newly introduced scope.  Otherwise, we
might incorrectly analyze code like this:

  const A = ...;
  @A // Resolves to the const above, not the type parameter below
  class C<A> {}

Previously, the way we handled most cases like this was as follows
(considering visitClassDeclaration as an example)

1. ResolverVisitor.visitClassDeclaration visited annotations first,
   before any new scopes had been created; this ensured that the
   annotations were resolved in the correct scope.

2. ResolverVisitor.visitClassDeclaration set the
   `_enclosingClassDeclaration` field to point to the ClassDeclaration
   node.

3. ResolverVisitor.visitClassDeclaration called
   ScopedVisitor.visitClassDeclaration, which created two new scopes
   for the class (one containing just type parameters, and one
   containing both type parameters and members); then it visited all
   the child nodes of the ClassDeclaration (including annotations)
   with one of those scopes active.  (This was incorrect, since
   neither of the two class scopes should apply to the annotations!)

4. When ResolverVisitor.visitAnnotation was reached, it noticed that
   the parent of the annotation matched `_enclosingClassDeclaration`,
   so it did an early return; this avoided overwriting the correct
   resolution produced in step 1.

This approach had some disadvantages:

- It meant that the scopes for annotations were only correct when
  using the ResolverVisitor.  The other class derived from
  ScopedVisitor (VariableResolverVisitor) got the scopes for
  annotations wrong.  This was benign, but it was getting in the way
  of my efforts to unify resolution logic between the analyzer and
  CFE.

- It was only properly implemented for ClassDeclaration,
  FunctionTypeAlias, and MixinDeclaration; annotations were
  incorrectly scoped for ClassTypeAlias, EnumDeclaration,
  ExtensionDeclaration, FunctionDeclaration,
  FunctionTypedFormalParameter, GenericTypeAlias, and
  MethodDeclaration.

This CL changes the approach so that ScopedVisitor always visits
annotations before creating scopes; this ensures that annotation
scoping is correct for both ResolverVisitor and
VariableResolverVisitor, and avoids the need for an early return from
the visitAnnotation method.

With this change, a few of the ScopedVisitor methods no longer wind up
making a super call into UnifyingAstVisitor (visitClassTypeAlias,
visitFunctionDeclarationInScope, visitFunctionTypeAliasInScope,
visitFunctionTypedFormalParameter, visitGenericTypeAlias, and
visitMethodDeclarationInScope); accordingly, I've updated the
ResolverVisitor to ensure that the elementResolver will still be
invoked for all these AST structures.  (Previously this happened
automatically because the UnifyingAstVisitor would call visitNode).

I've included several tests to ensure that we don't regress the
analyzer functionality that is fixed by this change.

Change-Id: I6c79ff8a12301e0300f0679acb489cce86f25955
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/200881
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>