[CFE] Fix name for dill extension builderes
If - for instance - reporting an error on a duplicate extensions with
the same name where (at least) one was from dill (and thus a dill
builder) we'd use the encoded name (e.g. `ExtensionName|get#actualName`)
instead of `actualName`, meaning that when we - as context - tries to
give a good error message, but creates a squiggly that's to long,
maybe saying something like

```
 org-dartlang-test:///lib1.dart:2:14: Context: This is one of the extension members.
   String get onObject => "object #1";
              ^^^^^^^^^^^^^^^^^^^^^
```

This CL fixes it by giving the expected name for
DillExtensionMemberBuilders.

Change-Id: I02ba460fb8ac8dbc081fdca485058608ae16e04f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/205265
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>